refactor(config): utilize `diable` option in the extract plugin to simplify config
Utilizing the `ExtractTextPlugin`s `disable` prop instead of duplicating the configs. Also
adding a `.postcssrc.js` which will make things easier going forward as we move to
postcss and cssnext instead of sass. This also removes some code from the webpack
config and isolates the postcss config which I think is a good thing.